REFACT: 用defineProperty重写扩展对象方法
