+++ src/extends/index.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="../typings/polyfill.d.ts" />
