+++ src/extends/index.ts
@@ -1,0 +1,4 @@
+import { HTMLElementPlus } from './HTMLElement';
+export default function nativeExtend() {
+  HTMLElementPlus();
+}
